added  support for `special foods`

+++ app/web/src/utils.rs
@@ -29,7 +29,7 @@ pub fn build_snake(positions: Vec<(u32, u32, Direction)>, max: &FieldPoint) -> S
     return snake;
 }
 
-pub fn to_base_10_array(n: u16, digits: u8) -> Vec<u8> {
+pub fn to_base_10_array(n: usize, digits: u8) -> Vec<u8> {
     let mut digits = vec![0; digits as usize];
     let mut n = n;
     let mut i = digits.len() - 1;
